update requirements. add test for transition pathway.
