+++ environment.yml
@@ -14,7 +14,6 @@ dependencies:
   - pip:
       - monty>=2.0.4
       - csdmpy>=0.3.4
-      - numexpr>=2.7.0
       - pydantic>=1.0
       - typing-extensions>=3.7
       - lmfit>=1.0.0
